Clean up codes
