+++ packages/shared/src/enterprise/index.ts
@@ -6,10 +6,7 @@ export * from "./power";
 export * from "./decision-criteria";
 export * from "./dashboards";
 export * from "./product-analytics/sql";
-export * from "./product-analytics/compare-date-range";
-export * from "./product-analytics/comparison-alignment";
-export * from "./product-analytics/compute-exploration-comparison";
-export * from "./product-analytics/densify-comparison-timeseries";
+export * from "./product-analytics/comparison";
 export * from "./product-analytics/utils";
 export * from "../validators/revisions";
 export * from "./validators/decision-criteria";
